Fix sequencing issues that prevent deterministic MPI runs
100+ seeded runs ranging from 2-6 chains each are reliably terminating and producing identical results.  Initial tests also show that gaussian_mpi_example.py is running twice as fast following this update, though larger-scale tests with different data take 1/3 longer (20 mins instead of 15 with 6+ chains).

1. Removed sequences of comm.Iprobe() + time.sleep() + comm.recv()

Use of time.sleep() appears from comments to have been a workaround to improve timing for the non-blocking Iprobe() call. This repeated sequence of calls seems to have caused non-deterministic parallel tempering runs, since they depend on unreliable timing of multiple MPI processes executing in parallel.  Timing of processes executing in parallel cannot be relied on without explicit synchronization (see next item).

2. Send messages for negatives as well as positives

Onstead of, on the sending side, sending only positive messages, then on the receiver side using Iprobe() [non-blocking] + sleep() + recv() [blocking], send messages for both positive and negative cases, always blocking on the receiver side until the determination is made on whether or not to proceed with the next step.  There may be some performance impacts here that need to be tested / optimized.

3. Fix a conditional error in PTSwap()

The second large conditional block only executed if the first didn't run, which meant that a lower temperature chain in the middle of a group would only communicate with higher temperature chains, but not respond to requests from lower temperature chains.  This created deadlocks during intermediate tests with > 2 chains.