MSTI feature added, temp replaced by beta
Model-Switch Thermodynamic Integration (MSTI) feature has been added, allowing users to sample across the posteriors of two models, instead of sampling between each model's prior and posterior, individually. Beta=0 is the posterior of model 0 and beta=1 is the posterior of model 1. A beta in between 0 and 1 is a mixture of the two models’ posterior distributions. Changes include the ability to give PTMCMC a tuple of prior and likelihood functions, a parameter of a tuple of lists of indices for each model’s parameters in MSTI used to sift through the combined set of parameters in p0, and a parameter to indicate the preferred ladder shape (only geometric or linear as of right now). When using MSTI, four additional columns are written in the chain file -- the log posterior and log likelihood of the two models being compared.  

Instances of 1/temperature have been replaced by beta. Changes include PTMCMC accepting beta max and beta min (while still accepting Tmin and Tmax), chain files named with beta instead of temperature, and a parameter to indicate whether the temperature naming convention should be use.

Temperature scaling has been commented out since for hot chains it produces massive and frequently invalid jumps.